Refactor: refactor sport profile dtos
