+++ src/main/kotlin/ar/edu/unq/ttip/sportbook/controller/SportProfileController.kt
@@ -1,68 +1,11 @@
 package ar.edu.unq.ttip.sportbook.controller
 
-import ar.edu.unq.ttip.sportbook.persistence.entity.Sport
+import ar.edu.unq.ttip.sportbook.dto.*
 import ar.edu.unq.ttip.sportbook.security.UserDetailsImpl
 import ar.edu.unq.ttip.sportbook.service.SportProfileService
 import org.springframework.http.ResponseEntity
 import org.springframework.security.core.annotation.AuthenticationPrincipal
 import org.springframework.web.bind.annotation.*
-
-data class UpdateFootballProfileRequest(
-    val positions: MutableList<String>? = mutableListOf(),
-    val favoritePosition: String? = null,
-    val ability: Int? = null,
-    val playsOften: Boolean? = null
-)
-
-data class FootballProfileDTO(
-    val sport: String = "FOOTBALL",
-    val positions: MutableList<String>? = mutableListOf(),
-    val favoritePosition: String? = null,
-    val ability: Int? = null,
-    val playsOften: Boolean? = null
-)
-
-data class UpdateVolleyProfileRequest(
-    val positions: MutableList<String>? = mutableListOf(),
-    val favoritePosition: String? = null,
-    val ability: Int? = null,
-    val playsOften: Boolean? = null,
-    val blockHeight: Int? = null,
-    val rolePreference: String? = null
-)
-
-data class VolleyProfileDTO(
-    val sport: String = "VOLLEY",
-    val positions: MutableList<String>? = mutableListOf(),
-    val favoritePosition: String? = null,
-    val ability: Int? = null,
-    val playsOften: Boolean? = null,
-    val blockHeight: Int? = null,
-    val rolePreference: String? = null
-)
-
-data class UpdatePaddleProfileRequest(
-    val preferredSide: String? = null,
-    val ability: Int? = null,
-    val playsOften: Boolean? = null,
-    val playStyle: String? = null,
-    val playedTournaments: Boolean? = null
-)
-
-data class PaddleProfileDTO(
-    val sport: String = "PADDLE",
-    val preferredSide: String? = null,
-    val ability: Int? = null,
-    val playsOften: Boolean? = null,
-    val playStyle: String? = null,
-    val playedTournaments: Boolean? = null
-)
-
-data class SportProfileDTO(
-    val sport: Sport,
-    val details: Any
-)
-
 
 @RestController
 @RequestMapping("/profile")
